Move finalize rule to base plugin
