+++ subprojects/platform-ios/src/main/java/dev/nokee/platform/ios/internal/IosApplicationComponentTag.java
@@ -1,0 +1,26 @@
+/*
+ * Copyright 2022 the original author or authors.
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *     https://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+package dev.nokee.platform.ios.internal;
+
+import dev.nokee.model.internal.core.ModelComponent;
+
+public final class IosApplicationComponentTag implements ModelComponent {
+	private static final IosApplicationComponentTag INSTANCE = new IosApplicationComponentTag();
+
+	public static IosApplicationComponentTag tag() {
+		return INSTANCE;
+	}
+}
