+++ subprojects/platform-ios/src/test/groovy/dev/nokee/platform/ios/ObjectiveCIosApplicationSourcesTest.java
@@ -21,8 +21,8 @@ import dev.nokee.language.objectivec.ObjectiveCSourceSet;
 import dev.nokee.language.objectivec.internal.plugins.ObjectiveCLanguageBasePlugin;
 import dev.nokee.model.internal.registry.ModelRegistry;
 import dev.nokee.platform.base.testers.ComponentSourcesTester;
+import dev.nokee.platform.ios.internal.plugins.IosComponentBasePlugin;
 import dev.nokee.platform.ios.internal.plugins.IosResourcePlugin;
-import dev.nokee.platform.nativebase.internal.plugins.NativeComponentBasePlugin;
 import lombok.val;
 
 import java.util.stream.Stream;
@@ -33,7 +33,7 @@ class ObjectiveCIosApplicationSourcesTest implements ComponentSourcesTester<Obje
 	@Override
 	public ObjectiveCIosApplicationSources createSubject() {
 		val project = ProjectTestUtils.rootProject();
-		project.getPluginManager().apply(NativeComponentBasePlugin.class);
+		project.getPluginManager().apply(IosComponentBasePlugin.class);
 		project.getPluginManager().apply(ObjectiveCLanguageBasePlugin.class);
 		project.getPluginManager().apply(IosResourcePlugin.class);
 		val sources = project.getExtensions().getByType(ModelRegistry.class).register(objectiveCIosApplication("main", project)).as(ObjectiveCIosApplication.class).get().getSources();
